// src/api/models/JournalDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type JournalDto = {
    id?: number;
    seq?: number;
    journalTypeId: number;
    journalTypeName?: string;
    createDate: string;
    filialId: number;
    filialName?: string;
    filialCode?: string;
    requesterName?: string;
    productName?: string;
    expertId: number;
    expertFirstName?: string;
    expertMiddleName?: string;
    expertLastName?: string;
    formId: number;
    formName?: string;
    formPrices?: string;
    expertizeNumber?: string;
    expertizeStartDate?: string;
    expertizeEndDate?: string;
    certificateNumber?: string;
    certificateBlankNumber?: number;
    receiverCountryId: number;
    receiverCountryShortName?: string;
    receiverCountryCode?: string;
    exportCountryId: number;
    exportCountryShortName?: string;
    exportCountryCode?: string;
    certificatePrice?: number;
    expertizePrice?: number;
    platformId: number;
    platformName?: string;
    receiverFio?: string;
    statusId?: number;
    statusName?: string;
    description?: string;
    productTypeId?: number;
    productTypeName?: string;
};