// src/api/models/JournalFilterDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type JournalFilterDto = {
    createDateFrom?: string;
    createDateTo?: string;
    changeDateTimeFrom?: string;
    changeDateTimeTo?: string;
    statusIds?: Array<number>;
    filialIds?: Array<number>;
    expertIds?: Array<number>;
    formIds?: Array<number>;
    certificatePrices?: Array<number>;
    productTypeIds?: Array<number>;
};