// src/api/models/PageableSortingDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PageableSortingDto = {
    sorting: string;
    sortingDirection: PageableSortingDto.sortingDirection;
};
export namespace PageableSortingDto {
    export enum sortingDirection {
        ASC = 'ASC',
        DESC = 'DESC',
    }
}